fix(polynomials.py): remove target_energies bug
the `calculate_complex_polynomial` method on `asp` classes used np.asarray(target_energies) which caused many problems downstream if `target_energies` was the default value of None. This has fixed that input to be the default energy values of the object.
in addition, changed variable modification to internal accessors for
factors, coefs and energies, such that warnings won't trigger on public
property accessors.