build(pyproject.toml): gui, precommit, semver
Add GUI scripts to target from commandline (i.e. kkcalc), add pre-commit
hooks by default with dev branch install, and defaulted matching of the
beta branch releases to anything that is not main. Also added a tkinter
error message for instances where __main__.py is called, but PyQT6 is not
installed.

Additionally added a fetch-depth 0 to semver, so reversion SHAs can be found.